+++ accumulo/src/main/java/org/vertexium/accumulo/util/DataInputStreamUtils.java
@@ -1,26 +1,26 @@
 package org.vertexium.accumulo.util;
 
-import com.google.common.base.Function;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Sets;
 import org.apache.hadoop.io.Text;
 import org.vertexium.FetchHints;
 import org.vertexium.FetchHints;
 import org.vertexium.Property;
 import org.vertexium.Visibility;
-import org.vertexium.accumulo.AccumuloGraph;
-import org.vertexium.accumulo.AccumuloNameSubstitutionStrategy;
-import org.vertexium.accumulo.LazyMutableProperty;
-import org.vertexium.accumulo.LazyPropertyMetadata;
+import org.vertexium.accumulo.*;
+import org.vertexium.accumulo.MetadataEntry;
 import org.vertexium.accumulo.iterator.model.*;
 import org.vertexium.accumulo.iterator.util.DataOutputStreamUtils;
 import org.vertexium.id.NameSubstitutionStrategy;
 
-import javax.annotation.Nullable;
 import java.io.DataInputStream;
 import java.io.IOException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import static org.vertexium.accumulo.iterator.util.DataInputStreamUtils.decodeIntArray;
 
 public class DataInputStreamUtils {
     public static Text decodeText(DataInputStream in) throws IOException {
@@ -29,18 +29,6 @@ public class DataInputStreamUtils {
 
     public static String decodeString(DataInputStream in) throws IOException {
         return org.vertexium.accumulo.iterator.util.DataInputStreamUtils.decodeString(in);
-    }
-
-    public static List<Text> decodeTextList(DataInputStream in) throws IOException {
-        int count = in.readInt();
-        if (count == -1) {
-            return null;
-        }
-        List<Text> results = new ArrayList<>();
-        for (int i = 0; i < count; i++) {
-            results.add(decodeText(in));
-        }
-        return results;
     }
 
     public static ImmutableSet<String> decodeStringSet(DataInputStream in) throws IOException {
@@ -55,9 +43,39 @@ public class DataInputStreamUtils {
         return results.build();
     }
 
+    public static List<MetadataEntry> decodeMetadataEntries(DataInputStream in) throws IOException {
+        int i = in.read();
+        if (i != ElementData.METADATA_START) {
+            throw new IOException(String.format("Unexpected metadata start: 0x%02x (expected: 0x%02x)", i, ElementData.METADATA_START));
+        }
+
+        int length = in.readInt();
+        List<MetadataEntry> results = new ArrayList<>(length);
+        for (i = 0; i < length; i++) {
+            results.add(decodeMetadataEntry(in));
+        }
+
+        i = in.read();
+        if (i != ElementData.METADATA_END) {
+            throw new IOException(String.format("Unexpected metadata end: 0x%02x (expected: 0x%02x)", i, ElementData.METADATA_END));
+        }
+
+        return results;
+    }
+
+    private static MetadataEntry decodeMetadataEntry(DataInputStream in) throws IOException {
+        String metadataKey = decodeString(in);
+        String metadataVisibility = decodeString(in);
+        int valueLength = in.readInt();
+        byte[] value = new byte[valueLength];
+        in.read(value);
+        return new MetadataEntry(metadataKey, metadataVisibility, value);
+    }
+
     public static Iterable<Property> decodeProperties(
             AccumuloGraph graph,
             DataInputStream in,
+            List<MetadataEntry> metadataEntries,
             FetchHints fetchHints
     ) throws IOException {
         List<Property> results = new ArrayList<>();
@@ -81,15 +99,11 @@ public class DataInputStreamUtils {
             Set<String> propertyHiddenVisibilitiesStringSet = decodeStringSet(in);
             Set<Visibility> propertyHiddenVisibilities = null;
             if (propertyHiddenVisibilitiesStringSet != null) {
-                propertyHiddenVisibilities = Sets.newHashSet(Iterables.transform(propertyHiddenVisibilitiesStringSet, new Function<String, Visibility>() {
-                    @Nullable
-                    @Override
-                    public Visibility apply(String input) {
-                        return new Visibility(input);
-                    }
-                }));
+                propertyHiddenVisibilities = propertyHiddenVisibilitiesStringSet.stream()
+                        .map(Visibility::new)
+                        .collect(Collectors.toSet());
             }
-            LazyPropertyMetadata metadata = decodePropertyMetadata(in, graph.getNameSubstitutionStrategy());
+            LazyPropertyMetadata metadata = decodePropertyMetadata(in, metadataEntries);
             results.add(new LazyMutableProperty(
                     graph,
                     graph.getVertexiumSerializer(),
@@ -106,21 +120,12 @@ public class DataInputStreamUtils {
         return results;
     }
 
-    private static LazyPropertyMetadata decodePropertyMetadata(DataInputStream in, AccumuloNameSubstitutionStrategy nameSubstitutionStrategy) throws IOException {
-        LazyPropertyMetadata metadata = new LazyPropertyMetadata();
-        int entryCount = in.readInt();
-        for (int i = 0; i < entryCount; i++) {
-            String key = nameSubstitutionStrategy.deflate(decodeString(in));
-            Visibility visibility = new Visibility(decodeString(in));
-            int valueLength = in.readInt();
-            byte[] value = new byte[valueLength];
-            int read = in.read(value);
-            if (read != valueLength) {
-                throw new IOException("Unexpected data length expected " + valueLength + " found " + read);
-            }
-            metadata.add(key, visibility, value);
-        }
-        return metadata;
+    private static LazyPropertyMetadata decodePropertyMetadata(
+            DataInputStream in,
+            List<MetadataEntry> metadataEntries
+    ) throws IOException {
+        int[] metadataIndexes = decodeIntArray(in);
+        return new IndexedLazyPropertyMetadata(metadataEntries, metadataIndexes);
     }
 
     public static Edges decodeEdges(DataInputStream in, NameSubstitutionStrategy nameSubstitutionStrategy) throws IOException {
@@ -160,10 +165,6 @@ public class DataInputStreamUtils {
             }
         }
         return edges;
-    }
-
-    public static byte[] decodeByteArray(DataInputStream in) throws IOException {
-        return org.vertexium.accumulo.iterator.util.DataInputStreamUtils.decodeByteArray(in);
     }
 
     public static void decodeHeader(DataInputStream in, byte expectedTypeId) throws IOException {